*** empty log message ***
git-svn-id: https://svn.nfit.au.dk/svn/beta/trunk@11730 678b6752-a8a0-49ff-a511-7c2a3241fcd4

+++ betarun/C/sockets.c
@@ -1160,7 +1160,7 @@ int closeSocket(int fd)
   if (closesocket(fd))
   {
     ERRNO = WSAGetLastError();
-    DEBUG_SOCKETS(printf("(Closesocket failed WSAGetLastError=%d)", ERRNO));
+    DEBUG_SOCKETS(printf("(Closesocket failed WSAGetLastError=%d)", (int)ERRNO));
     if (ERRNO == WSAEWOULDBLOCK)
       return 1;
     return -1;
